chore: export factory from spacing
